use size_t to reduce warnings

+++ src/luvit_exports.c
@@ -33,27 +33,27 @@ extern const char **luaJIT_BC_utils;
 const void *suck_in_luvit(void)
 {
   luvit_ugly_hack = (const void*)
-    (int)(const void *)luaJIT_BC_dns +
-    (int)(const void *)luaJIT_BC_emitter +
-    (int)(const void *)luaJIT_BC_fiber +
-    (int)(const void *)luaJIT_BC_fs +
-    (int)(const void *)luaJIT_BC_http +
-    (int)(const void *)luaJIT_BC_luvit +
-    (int)(const void *)luaJIT_BC_mime +
-    (int)(const void *)luaJIT_BC_net +
-    (int)(const void *)luaJIT_BC_path +
-    (int)(const void *)luaJIT_BC_pipe +
-    (int)(const void *)luaJIT_BC_process +
-    (int)(const void *)luaJIT_BC_repl +
-    (int)(const void *)luaJIT_BC_request +
-    (int)(const void *)luaJIT_BC_response +
-    (int)(const void *)luaJIT_BC_stack +
-    (int)(const void *)luaJIT_BC_stream +
-    (int)(const void *)luaJIT_BC_tcp +
-    (int)(const void *)luaJIT_BC_tty +
-    (int)(const void *)luaJIT_BC_udp +
-    (int)(const void *)luaJIT_BC_url +
-    (int)(const void *)luaJIT_BC_utils;
+    (size_t)(const void *)luaJIT_BC_dns +
+    (size_t)(const void *)luaJIT_BC_emitter +
+    (size_t)(const void *)luaJIT_BC_fiber +
+    (size_t)(const void *)luaJIT_BC_fs +
+    (size_t)(const void *)luaJIT_BC_http +
+    (size_t)(const void *)luaJIT_BC_luvit +
+    (size_t)(const void *)luaJIT_BC_mime +
+    (size_t)(const void *)luaJIT_BC_net +
+    (size_t)(const void *)luaJIT_BC_path +
+    (size_t)(const void *)luaJIT_BC_pipe +
+    (size_t)(const void *)luaJIT_BC_process +
+    (size_t)(const void *)luaJIT_BC_repl +
+    (size_t)(const void *)luaJIT_BC_request +
+    (size_t)(const void *)luaJIT_BC_response +
+    (size_t)(const void *)luaJIT_BC_stack +
+    (size_t)(const void *)luaJIT_BC_stream +
+    (size_t)(const void *)luaJIT_BC_tcp +
+    (size_t)(const void *)luaJIT_BC_tty +
+    (size_t)(const void *)luaJIT_BC_udp +
+    (size_t)(const void *)luaJIT_BC_url +
+    (size_t)(const void *)luaJIT_BC_utils;
 
   return luvit_ugly_hack;
 }
